+++ src/OrchardCore/OrchardCore.Infrastructure.Abstractions/Security/Services/IPermissionsService.cs
@@ -11,32 +11,4 @@ namespace OrchardCore.Security.Services
     {
         Task<IEnumerable<Permission>> GetInstalledPermissionsAsync();
     }
-
-    public class BasePermissionsService : IPermissionsService
-    {
-
-        private readonly IEnumerable<IPermissionProvider> _permissionProviders;
-
-        public BasePermissionsService(IEnumerable<IPermissionProvider> permissionProviders)
-        {
-            _permissionProviders = permissionProviders;
-        }
-
-        public async Task<IEnumerable<Permission>> GetInstalledPermissionsAsync()
-        {
-            var installedPermissions = new List<Permission>();
-
-            foreach (var permissionProvider in _permissionProviders)
-            {
-                var permissions = await permissionProvider.GetPermissionsAsync();
-
-                foreach (var permission in permissions)
-                {
-                    installedPermissions.Add(permission);
-                }
-            }
-
-            return installedPermissions;
-        }
-    }
 }
